Ignore events that are too close to each other.
When using other software that either remaps or sends `CGEvent` in rapid
succession, the sounds from Thock get muffled out and overlap. To avoid
this, we do a tinsy bit of booking of when the previous event was sent
and check if the next event has enough elapsed time to be a valid event.